Early logging configuration (#1734)
